Moved card interfaces and types into shared module.
