+++ server/Shared Models/src/RouteType.kt
@@ -1,0 +1,13 @@
+import java.io.Serializable
+
+enum class RouteType(val type: String) : Serializable {
+    ANY("any"),
+    REALITY("reality"),
+    SOUL("soul"),
+    SPACE("space"),
+    MIND("mind"),
+    POWER("power"),
+    TIME("time"),
+    VIBRANIUM("vibranium"),
+    PALLADIUM("palladium")
+}
