+++ server/Thanos-AI/src/Thanos.kt
@@ -1,6 +1,6 @@
 /**
  * Created by Jordan Gassaway on 4/22/2019.
- * Thanos TODO: Enter Class Purpose Here
+ * Thanos - Self contained AI that can play the ticket to ride game
  */
 class Thanos {
 }
